requirement.txt
a2wsgi==1.10.10
agate==1.9.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
aiosmtplib==5.0.0
aiosqlite==0.21.0
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
apache-airflow==3.1.0
apache-airflow-core==3.1.0
apache-airflow-providers-amazon==9.15.0
apache-airflow-providers-celery==3.12.3
apache-airflow-providers-common-compat==1.7.4
apache-airflow-providers-common-io==1.6.3
apache-airflow-providers-common-sql==1.28.1
apache-airflow-providers-http==5.3.4
apache-airflow-providers-smtp==2.2.1
apache-airflow-providers-snowflake==6.5.4
apache-airflow-providers-standard==1.8.0
apache-airflow-task-sdk==1.1.0
apispec==6.8.4
argcomplete==3.6.2
asgiref==3.9.2
asn1crypto==1.5.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.14.2
billiard==4.2.2
blinker==1.9.0
boto3==1.40.55
botocore==1.40.55
cachelib==0.13.0
cachetools==6.2.1
cadwyn==5.4.4
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
clickclick==20.10.2
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.0
daff==1.4.2
dbt-adapters==1.17.3
dbt-common==1.32.0
dbt-core==1.10.13
dbt-extractor==0.6.0
dbt-protos==1.0.382
dbt-semantic-interfaces==0.9.0
dbt-snowflake==1.10.2
deepdiff==8.6.1
Deprecated==1.2.18
dill==0.3.1.1
dnspython==2.8.0
durationpy==0.10
email-validator==2.3.0
fastapi==0.117.1
fastapi-cli==0.0.13
filelock==3.20.0
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
flower==2.0.1
frozenlist==1.8.0
fsspec==2025.9.0
google-auth==2.41.1
google-re2==1.1.20250805
googleapis-common-protos==1.70.0
greenback==1.2.1
greenlet==3.2.4
grpcio==1.65.5
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
humanize==4.13.0
idna==3.10
importlib_metadata==8.4.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
jsonpath-ng==1.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.6.0
kombu==5.5.4
kubernetes==33.1.0
kubernetes_asyncio==33.3.0
lazy-object-proxy==1.12.0
leather==0.4.0
libcst==1.8.4
limits==5.6.0
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.4.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.28.2
mashumaro==3.14
mdit-py-plugins==0.5.0
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
natsort==8.4.0
networkx==3.5
numpy==2.3.4
oauthlib==3.3.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
ordered-set==4.1.0
orderly-set==5.5.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
parsedatetime==2.6
pathspec==0.12.1
pendulum==3.1.0
platformdirs==4.5.0
pluggy==1.6.0
ply==3.11
prison==0.2.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.31.1
psutil==7.1.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAthena==3.19.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-nvd3==0.16.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
redshift-connector==2.1.9
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retryhttp==1.3.3
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.1
rich-toolkit==0.15.1
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
sagemaker_studio==1.0.22
scramp==1.4.6
SecretStorage==3.4.0
setproctitle==1.3.7
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.18.0
snowflake-snowpark-python==1.41.0
snowflake-sqlalchemy==1.7.7
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.0
sqlparse==0.5.3
starlette==0.48.0
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
tomlkit==0.13.3
tornado==6.5.2
typer==0.19.2
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.37.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
watchtower==3.4.0
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==2.2.3
wheel==0.45.1
wirerope==1.0.0
wrapt==1.17.3
WTForms==3.2.1
yarl==1.22.0
zipp==3.23.0
